refactor(ui): move the network selected types
