+++ ui/src/app/base/components/node/networking/types.ts
