+++ ui/src/app/machines/views/MachineDetails/MachineNetwork/EditBondForm/EditBondForm.tsx
@@ -15,9 +15,12 @@ import {
 } from "../BondForm/utils";
 import InterfaceFormTable from "../InterfaceFormTable";
 import { networkFieldsSchema } from "../NetworkFields/NetworkFields";
-import type { Selected, SetSelected } from "../NetworkTable/types";
 
 import FormikForm from "app/base/components/FormikForm";
+import type {
+  Selected,
+  SetSelected,
+} from "app/base/components/node/networking/types";
 import { useIsAllNetworkingDisabled } from "app/base/hooks";
 import { MAC_ADDRESS_REGEX } from "app/base/validation";
 import { useMachineDetailsForm } from "app/machines/hooks";
